// src/types.ts
export interface ParsedPAData {
    ownerName?: string;
    mailingAddress?: string;
    folio?: string;
    legalDescription?: string;
    street?: string;
    city?: string;
    zip?: string;
    siteAddress?: string;
    imageUrl?: string;
    sketchPath?: string;
    parcelPhotoPath?: string | null;
  }